+++ configuration.txt
@@ -1,2 +1,3 @@
 1. npm install
 2. npm install --save react-router-dom
+3. npm install react-router-dom
